+++ src/solver_lbm/solver_lbm_kernels.cu
@@ -26,7 +26,7 @@ void Cu_SetInitialConditions
 {
     constexpr int N_DIM = AP->N_DIM;
     constexpr int M_CBLOCK = AP->M_CBLOCK;
-    constexpr int N_Q = LP->N_Q;
+    constexpr long int N_Q = LP->N_Q;
     int I = threadIdx.x % 4;
     int J = (threadIdx.x / 4) % 4;
     int K = 0;
@@ -119,7 +119,7 @@ void Cu_RefreshVariables
 {
     constexpr int N_DIM = AP->N_DIM;
     constexpr int M_CBLOCK = AP->M_CBLOCK;
-    constexpr int N_Q = LP->N_Q;
+    constexpr long int N_Q = LP->N_Q;
     
     int i_kap_b = -1;
     if (blockIdx.x < n_ids_idev_L)
@@ -219,7 +219,7 @@ void Cu_Collide
 {
     constexpr int N_DIM = AP->N_DIM;
     constexpr int M_CBLOCK = AP->M_CBLOCK;
-    constexpr int N_Q = LP->N_Q;
+    constexpr long int N_Q = LP->N_Q;
     
     int i_kap_b = -1;
     if (blockIdx.x < n_ids_idev_L)
@@ -306,7 +306,7 @@ void Cu_CollideLES
 {
     constexpr int N_DIM = AP->N_DIM;
     constexpr int M_CBLOCK = AP->M_CBLOCK;
-    constexpr int N_Q = LP->N_Q;
+    constexpr long int N_Q = LP->N_Q;
     
     int i_kap_b = -1;
     if (blockIdx.x < n_ids_idev_L)
@@ -399,7 +399,7 @@ void Cu_CollideLES
 {
     constexpr int N_DIM = AP->N_DIM;
     constexpr int M_CBLOCK = AP->M_CBLOCK;
-    constexpr int N_Q = LP->N_Q;
+    constexpr long int N_Q = LP->N_Q;
     
     int i_kap_b = -1;
     if (blockIdx.x < n_ids_idev_L)
@@ -551,7 +551,7 @@ void Cu_Stream
     constexpr int N_DIM = AP->N_DIM;
     constexpr int N_Q_max = AP->N_Q_max;
     constexpr int M_CBLOCK = AP->M_CBLOCK;
-    constexpr int N_Q = LP->N_Q;
+    constexpr long int N_Q = LP->N_Q;
     __shared__ int s_ID_nbr[27];
     
     int i_kap_b = -1;
@@ -712,7 +712,7 @@ void Cu_ImposeBC
     constexpr int N_DIM = AP->N_DIM;
     constexpr int M_CBLOCK = AP->M_CBLOCK;
     constexpr int N_Q_max = AP->N_Q_max;
-    constexpr int N_Q = LP->N_Q;
+    constexpr long int N_Q = LP->N_Q;
     __shared__ int s_ID_nbr[N_Q_max];
     int I = threadIdx.x % 4;
     int J = (threadIdx.x / 4) % 4;
